+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/services/ServicesModule.kt
@@ -1,8 +1,8 @@
 package org.noiseplanet.noisecapture.services
 
 import org.koin.dsl.module
-import org.noiseplanet.noisecapture.services.liveaudio.DefaultLiveAudioService
-import org.noiseplanet.noisecapture.services.liveaudio.LiveAudioService
+import org.noiseplanet.noisecapture.services.audio.DefaultLiveAudioService
+import org.noiseplanet.noisecapture.services.audio.LiveAudioService
 import org.noiseplanet.noisecapture.services.location.DefaultUserLocationService
 import org.noiseplanet.noisecapture.services.location.UserLocationService
 import org.noiseplanet.noisecapture.services.measurement.DefaultMeasurementRecordingService
@@ -45,6 +45,7 @@ val servicesModule = module {
             measurementService = get(),
             liveAudioService = get(),
             userLocationService = get(),
+            audioRecordingService = get(),
         )
     }
 }
